+++ test_app/urls.py
@@ -1,5 +1,12 @@
 from mapentity.registry import registry
-from test_app.models import City, DummyAptModel, DummyModel, MushroomSpot, Road
+from test_app.models import (
+    City,
+    DummyAptModel,
+    DummyModel,
+    GeoPoint,
+    MushroomSpot,
+    Road,
+)
 
 app_name = "test_app"
 
@@ -8,3 +15,4 @@ urlpatterns += registry.register(MushroomSpot)
 urlpatterns += registry.register(Road)
 urlpatterns += registry.register(City)
 urlpatterns += registry.register(DummyAptModel)
+urlpatterns += registry.register(GeoPoint)